# Erronka2/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "APIs For Thrift Store",
        "version": "1.0.0"
    },
    "paths": {
...
        "/talde2erronka2back/Erronka2/public/api/materialaF/{id}": {
            "get": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Jakin ahal izateko material bat libre edo okupaturik dagoen",
                "operationId": "8a3a04f80a002bcc65395b7f43a62e2c",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Materialaren IDa",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Materialen datuak lortu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/materiala/devolver": {
            "put": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Materiala bueltatzeko funtzioa",
                "operationId": "19c4b593c1be8d399769cf67b31a7ab3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_materiala"
                                ],
                                "properties": {
                                    "id_materiala": {
                                        "description": "Materialaren IDa",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materiala bueltatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/materiala/ezabatu": {
            "put": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Materiala ezabatzeko funtzioa",
                "operationId": "f01bd11d031d7babd0ad1341d1036fb3",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Materialaren IDa",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materiala ezabatu da."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/materiala/txertatu": {
            "post": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Material berria sartzeko funtzioa",
                "operationId": "80c24c20e5c5b2ee0d03a432b175789b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "etiketa",
                                    "izena"
                                ],
                                "properties": {
                                    "etiketa": {
                                        "description": "Materialaren etiketa",
                                        "type": "string",
                                        "example": "example_etiketa"
                                    },
                                    "izena": {
                                        "description": "Materialaren izena",
                                        "type": "string",
                                        "example": "example_izena"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materiala txertatu da."
                    }
                }
            }
        },
...
        "/talde2erronka2back/Erronka2/public/api/materiala/alumnos/{kodea}": {
            "get": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Erreserba egiterakoan zein ikasle dauden jakiteko",
                "operationId": "7cb12e9ea2d7063a1a34efdb55ffe2f0",
                "parameters": [
                    {
                        "name": "kodea",
                        "in": "path",
                        "description": "Taldearen kodea",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Datuak lortu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/materiala/reservar": {
            "post": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Erreserba egiterakoan zein ikasle dauden jakiteko",
                "operationId": "964a3bb8033d34ee70f7e601bbfb0ab0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id_materiala",
                                    "id_langilea"
                                ],
                                "properties": {
                                    "id_materiala": {
                                        "description": "Materialaren IDa",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "id_langilea": {
                                        "description": "Langilearen IDa",
                                        "type": "integer",
                                        "example": "1"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Datuak lortu dira."
                    }
                }
            }
        },
        "/talde2erronka2back/Erronka2/public/api/materiala/editatu": {
            "put": {
                "tags": [
                    "Materiala"
                ],
                "summary": "Materiala editatu ahal izateko funtzioa",
                "operationId": "36c1f8a1fb69372d0c6c5d97c08d1cbf",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "id",
                                    "etiketa",
                                    "izena"
                                ],
                                "properties": {
                                    "id": {
                                        "description": "Materialaren IDa",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "etiketa": {
                                        "description": "Materialaren etiketa",
                                        "type": "string",
                                        "example": "example_etiketa"
                                    },
                                    "izena": {
                                        "description": "Materialaren izena",
                                        "type": "string",
                                        "example": "example_izena"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Materiala editatu da."
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}